Two-dimensional Clayton Copula

This is the joint cumulative distribution and probability density functions for the two-dimensional Clayton copula.

Usage

Clayton_Copula(u, v, theta)

Clayton_Copula_pdf(u, v, theta)

Arguments

u

is a vector of points in \([0,1]\) representing the first coordinate of the copula.

v

is a vector of points in \([0,1]\) representing the first coordinate of the copula.

theta

defines \(\theta\) dependency parameter between the pairs.

Value

Clayton_Copula gives the joint cumulative distribution function, Clayton_Copula_pdf gives the joint probability density function values for the Clayton copula evaluated at points \((u,v)\) with given \(\theta\) parameter.

Details

The joint distribution function of the two-dimensional Clayton copula, along with its joint probability density function, are given by $$C(u,v;\theta) = \left(u^{-\theta} + v^{-\theta} - 1\right)^{-1/\theta}, \; \theta \in (0, \infty).$$ $$c_{\theta}(u,v) = (\theta +1) u^{-(\theta + 1)} v^{-(\theta + 1)} \left( u^{-\theta} + v^{-\theta}-1 \right)^{-\left (\frac{1}{\theta} + 2 \right)}$$ where \(\theta > 0\).