The Modified Weibull Distribution (MWD)

This is the probability density function, cumulative distribution function, quantile function, random generation and hazard function for the Modified Weibull distribution (MWD) by Lai et al. (2003).

Usage

dMweibull(x, a, b, lambda, log = FALSE)

pMweibull(q, a, b, lambda, lower.tail = TRUE, log = FALSE)

qMweibull(p, a, b, lambda, lower.tail = TRUE)

rMweibull(n, a, b, lambda)

hMweibull(x, a, b, lambda, log = FALSE)

Arguments

x

Numeric vector of observations.

a

Positive scale parameter of the distribution (\(a > 0\)).

b

Non-negative shape parameter of the distribution (\(b \ge 0\)).

lambda

Non-negative acceleration or flexibility parameter (\(\lambda \ge 0\)) that governs how quickly the risk of failure (hazard) grows over time. Larger values correspond to faster failure accumulation and higher fragility, and restricting \(\lambda \ge 0\) ensures a non-decreasing hazard function.

log

Logical; if TRUE, then \(log(f(t))\) or \(log(F(t))\) or \(log(h(t))\) are returned.

q

Numeric vector of quantiles.

lower.tail

Logical; if FALSE, then \(1-F(t)\) are returned and quantiles are computed for \(1-p\).

p

Numeric vector of probabilities in \([0,1]\).

n

Integer; number of observations to be generated.

Value

dMweibull gives the probability density function, pMweibull gives the cumulative distribution function, qMweibull gives the quantile function, rMweibull generates random deviates, and hMweibull gives the hazard function.

Details

The Modified Weibull distribution with parameters \(a\), \(b\) and \(\lambda\) has the cumulative distribution density and hazard functions given by $$F(x) = 1 - \exp(-a x^b \exp(\lambda x)),$$ $$f(x) = a (b+ \lambda x) x^{b - 1} \exp(\lambda x) \exp(-a x^b \exp(\lambda x)),$$ $$h(x) = a (b+ \lambda x) x^{b - 1} \exp(\lambda x), $$ where \(x > 0\), \(a > 0\) is the scale parameter, \(b \ge 0\) is a shape parameter, and \(\lambda \ge 0\) is an acceleration or flexibility parameter that controls how quickly the hazard grows over time. When \(\lambda = 0\), the MWD reduces to the two-parameter Weibull distribution \(F(x) = 1 - \exp(-a x^b)\) with the scale parameter \(a\) and the shape parameter \(b\). When \(b = 0\), the MWD reduces to a type I extreme-value (or log-gamma) distribution with \(F(x) = 1 - \exp(-a \exp(\lambda x)) \).

References

Lai, C. D., Xie, M., and Murthy, D. N. P. (2003). https://doi.org/10.1109/TR.2002.805788 IEEE Transactions on Reliability, 52(1), 33–37.

Examples

n <- 25
a <- 0.75; b <- 1.25; lambda <- 0.60
set.seed(123)
x <- rlnorm(n)
ff <- dMweibull(x, a, b, lambda)
FF <- pMweibull(x, a, b, lambda)
qq <- qMweibull(runif(n), a, b, lambda)
dat <- rMweibull(n, a, b, lambda)
hf <- hMweibull(x, a, b, lambda )