Stress-Strength Reliability of MWD under Clayton copula

This is the main function to evaluate the stress-strength reliability (SSR) \(R=P(X>Y)\) where \(X \sim MWD(a_1,b_1,\lambda_1)\) strength and \(Y \sim MWD(a_2,b_2,\lambda_2)\) stress variables follow the MWD when the dependency is constructed by Clayton copula.

Reliability_Clayton_MWD(a1, b1, lambda1, a2, b2, lambda2, theta)

Arguments

a1, b1, lambda1

MWD parameters for the strength variable \(X\), with \(a_1 > 0\), \(b_1 \ge 0\), and \(\lambda_1 \ge 0\).

a2, b2, lambda2

MWD parameters for the stress variable \(Y\), with \(a_2 > 0\), \(b_2 \ge 0\), and \(\lambda_2 \ge 0\).

theta

Clayton copula parameter with \(\theta >0\).

Value

It is the same as a value of integrate function in stats.

value

The final estimate of the integral.

abs.error

Estimate of the modulus of the absolute error.

Details

The SSR \(R\) is given by $$ R = \int_{0}^{\infty} F_X(x)^{-(\theta + 1)} \left( F_X(x)^{-\theta} + G_Y(x)^{-\theta} - 1 \right)^{-\left(\frac{1}{\theta}+1\right)} f_X(x)\, dx, $$ which can also be written as $$ R = \int_{0}^{1} t^{-(\theta + 1)} \left( t^{-\theta} + G_Y(F_X^{-1}(t))^{-\theta} - 1 \right)^{-\left(\frac{1}{\theta}+1\right)} dt, $$

where \(F_X(x) = 1 - \exp(-a_1 x^{b_1} e^{\lambda_1 x})\) and \(G_Y(y) = 1 - \exp(-a_2 y^{b_2} e^{\lambda_2 y})\). Further details can be found in Kızılaslan (2026).

References

Kızılaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv:????

Examples

a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5

R = Reliability_Clayton_MWD(a1, b1, lambda1, a2, b2, lambda2, theta)
R$value
#> [1] 0.8805338

R_MC = Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 100000)
R_MC$R
#> [1] 0.8802808