Monte Carlo Estimation of Stress–Strength Reliability (MWD + Clayton Copula)

Computes an approximate value of the stress–strength reliability \(R = P(X > Y)\) using Monte Carlo integration method under the modified Weibull model with dependence induced by a Clayton copula.

It calculates an approximate value of \(R\) using the Monte Carlo integration method based on a random generated MWD sample from \(MWD(a_1, b_1, \lambda_1)\).

Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 10000)

Arguments

a1, b1, lambda1

Parameters of the strength variable \(X\), with \(a_1 > 0\), \(b_1 \ge 0\), and \(\lambda_1 \ge 0\).

a2, b2, lambda2

Parameters of the stress variable \(Y\), with \(a_2 > 0\), \(b_2 \ge 0\), and \(\lambda_2 \ge 0\).

theta

Clayton copula dependence parameter, \(\theta > 0\).

N

Integer. Number of Monte Carlo samples from \(X \sim MWD(a_1,b_1,\lambda_1)\).

Value

A list containing:

R

Approximated reliability value based on \(N\) generated random samples.

R_MCsample

Monte Carlo approximate of the reliability \(R\) based on \(N\) generated samples.

sample

Generated random sample from \(MWD(a_1, b_1, \lambda_1)\) with \(N\) size.

Details

Monte Carlo Estimation of Stress–Strength Reliability under Clayton Copula

The approximate stress–strength reliability \(R\) is approximated via Monte Carlo integration:

$$ R \approx \frac{1}{N} \sum_{i=1}^{N} T(x_i; \boldsymbol{\Omega_1}, \boldsymbol{\Omega_2}, \theta) = \tilde{R}, $$

where \(\boldsymbol{\Omega_1} = (a_1, b_1, \lambda_1)\) and \(\boldsymbol{\Omega_2} = (a_2, b_2, \lambda_2)\) denote the marginal parameter vectors. The function

$$ T(x; \boldsymbol{\Omega_1}, \boldsymbol{\Omega_2}, \theta) = F_X(x)^{-(\theta + 1)} \left( F_X(x)^{-\theta} + G_Y(x)^{-\theta} - 1 \right)^{-\left(\frac{1}{\theta}+1\right)}, $$

where \(x_1, \dots, x_N\) is a random sample generated from the Modified Weibull distribution \(MWD(a_1, b_1, \lambda_1)\).

The accuracy of the approximation improves as sample size \(N\) increases. In practice, values around \(N = 10^5\) typically provide stable results.

Further details can be found in Kizilaslan (2026).

References

Kizilaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv:2604.12130

Examples

# example code
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5
R <- Reliability_Clayton_MWD(a1, b1, lambda1, a2, b2, lambda2, theta)
R$value
#> [1] 0.8805338
# approximated reliability R based on MC method
R_MC <- Reliability_Clayton_MWD_MC(a1, b1, lambda1, a2, b2, lambda2, theta, N = 100000)
R_MC$R
#> [1] 0.8806282