One-step WLSE estimate of the Clayton copula parameter

One-step WLSE estimate of the Clayton copula parameter

Usage

WLSE_clayton_onestep(par, x, y, estimates)

Arguments

par

Numeric value of the dependence parameter \(\theta\) of the Clayton copula.

x

Vector of observations for the strength variable \(X\).

y

Vector of observations for the stress variable \(Y\).

estimates

A list containing estimates of the model parameters \(a_1, b_1, \lambda_1, a_2, b_2, \lambda_2\).

Value

The one-step WLSE of \(\theta\) parameter.

Details

Further details are provided in Kızılaslan (2026).

References

Kızılaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv:????