Estimating parameter of the Clayton copula

Estimates the dependence parameter \(\theta\) of the Clayton copula.

Usage

fitClayton(
  x,
  y,
  est.method,
  opt.method,
  start,
  estimates,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)

Arguments

x

Vector of observations for the strength variable \(X\).

y

Vector of observations for the stress variable \(Y\).

est.method

Used method for estimating the parameters such as maximum likelihood estimate "MLE", the least square estimation "LSE", the weighted least square estimation "WLSE", and the maximum product of spacing estimates "MPS".

opt.method

The optimization method for optim function such as "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.

start

Initial value for \(\theta\) parameter to be optimized over.

estimates

A list containing estimates of all model paramters \(a_1,b_1,\lambda_1, a_2, b_2, \lambda_2\).

lower

Numeric vector specifying the lower bounds of the parameters for bounded optimization methods (e.g., "L-BFGS-B"). Ignored if NULL.

upper

Numeric vector specifying the upper bounds of the parameters for bounded optimization methods. Ignored if NULL.

verbose

Logical; if TRUE, progress and intermediate results from the optimization procedure are printed. Default is FALSE.

...

Further arguments to be passed to optim function such as lower and upper limits, hessian, etc.

Value

A list containing:

estimate

Estimate of the Clayton copula parameter, \(\theta\).

opt.fit

Full optimization result.

Details

The joint distribution function of the two-dimensional Clayton copula is $$ C(u,v;\theta) = \left(u^{-\theta} + v^{-\theta} - 1\right)^{-1/\theta}, $$ where the dependence parameter satisfies \(\theta > 0\). It is estimated using several classical methods:

Maximum Likelihood Estimation (MLE): Obtains parameter estimates by maximizing the log-likelihood function of the observed data under the assumed MWD distribution.

Least Squares Estimation (LSE): Estimates parameters by minimizing the sum of squared differences between the empirical distribution function and theoretical distribution function of MWD. The Benard’s approximation is used for the empirical distribution function at the ordered observations as \(F(x_(i)) = (i-0.3)/(n+0.4), i=1,...,n.\).

Weighted Least Squares Estimation (WLSE): A modification of LSE that assigns weights to the squared differences. The weights are taken as \(w_i = ((n+1)^2(n+2)) / (i(n-i+1)), i=1,...,n. \)

Maximum Product of Spacings (MPS): Estimates parameters by maximizing the product of spacings between consecutive values of the fitted distribution function, providing a robust alternative to MLE, particularly in small samples.

Further details can be found in Kızılaslan (2026).

References

Kızılaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv preprint. Available at https://arxiv.org/abs/????.