Estimating parameters of the Modified Weibull Distribution (MWD)

Estimates the parameters of the Modified Weibull Distribution (MWD) using classical methods.

Usage

fitMWD(
  data,
  est.method,
  opt.method,
  starts,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)

Arguments

data

Vector of observations.

est.method

Used method for estimating the parameters such as maximum likelihood estimate "MLE", the least square estimation "LSE", the weighted least square estimation "WLSE", and the maximum product of spacing estimates "MPS".

opt.method

The optimization method for optim function such as "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.

starts

Initial values for the parameters to be optimized over.

lower

Numeric vector specifying the lower bounds of the parameters for bounded optimization methods (e.g., "L-BFGS-B"). Ignored if NULL.

upper

Numeric vector specifying the upper bounds of the parameters for bounded optimization methods. Ignored if NULL.

verbose

Logical; if TRUE, progress and intermediate results from the optimization procedure are printed. Default is FALSE.

...

Further arguments to be passed to optim function such as lower and upper limits, hessian, etc.

Value

A list containing:

estimates

Estimated values of the model parameters.

measures

Model selection criteria, including the log-likelihood, AIC, and BIC, evaluated at the estimated parameter values.

initials

Initial values used in the optimization procedure.

opt.fit

Full optimization output.

Details

Classical estimations for the parameters of the Modified Weibull Distribution (MWD)

The Modified Weibull Distribution (Lai et al., 2003) has cumulative distribution function (CDF) and probability density function (PDF) given by

$$ F(x) = 1 - \exp\left(-a x^b \exp(\lambda x)\right), $$ $$ f(x) = a (b + \lambda x) x^{b - 1} \exp(\lambda x) \exp\left(-a x^b \exp(\lambda x)\right), $$

where \(x > 0\), \(a > 0\) is a scale parameter, \(b \ge 0\) is a shape parameter, and \(\lambda \ge 0\) is a flexibility parameter controlling the growth rate of the hazard function.

The model parameters are estimated using several classical methods:

Maximum Likelihood Estimation (MLE): Obtains parameter estimates by maximizing the log-likelihood function of the observed data under the assumed MWD distribution.

Least Squares Estimation (LSE): Estimates parameters by minimizing the sum of squared differences between the empirical distribution function and theoretical distribution function of MWD. The Benard’s approximation is used for the empirical distribution function at the ordered observations as \(F(x_(i)) = (i-0.3)/(n+0.4), i=1,...,n.\).

Weighted Least Squares Estimation (WLSE): A modification of LSE that assigns weights to the squared differences. The weights are taken as \(w_i = ((n+1)^2(n+2)) / (i(n-i+1)), i=1,...,n. \)

Maximum Product of Spacings (MPS): Estimates parameters by maximizing the product of spacings between consecutive values of the fitted distribution function, providing a robust alternative to MLE, particularly in small samples.

Further details can be found in Kızılaslan (2026).

References

Lai, C. D., Xie, M., and Murthy, D. N. P. (2003). https://doi.org/10.1109/TR.2002.805788 IEEE Transactions on Reliability, 52(1), 33–37.

Kızılaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv preprint. Available at https://arxiv.org/abs/????.

Examples

# generate data from MWD(a, b, lambda)
n <- 100
a <- 0.75; b <- 1.25; lambda <- 0.60
set.seed(123)
dat <- rMweibull(n, a, b, lambda)
init <- runif(3)

# Fit MWD to dat.
fit.mle <- fitMWD(data = dat, est.method = "mle", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.mle$estimates
#>         a         b    lambda 
#> 0.7231634 1.2600843 0.6559157 

fit.lse <- fitMWD(data = dat, est.method = "lse", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.lse$estimates
#>         a         b    lambda 
#> 0.9299033 1.4069386 0.4020883 

fit.wlse <- fitMWD(data = dat, est.method = "wlse", opt.method = "L-BFGS-B", starts = init,
                   lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.wlse$estimates
#>         a         b    lambda 
#> 0.9220048 1.4228131 0.4207337 

fit.mps <- fitMWD(data = dat, est.method = "mps", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05,1e-05), upper = c(Inf,Inf,Inf), hessian = FALSE )
fit.mps$estimates
#>        a        b   lambda 
#> 0.714056 1.189107 0.646072