Bivariate Random Data Generation for MWD Marginals via Clayton Copula

Generates bivariate random samples from a dependent stress–strength model where both marginals follow the Modified Weibull Distribution (MWD), and the dependence structure between the variables is modeled using a Clayton copula.

Usage

rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)

Arguments

n

Integer; number of observations to be generated.

a1, b1, lambda1

MWD parameters for the strength variable \(X\), with \(a_1 > 0\), \(b_1 \ge 0\), and \(\lambda_1 \ge 0\).

a2, b2, lambda2

MWD parameters for the stress variable \(Y\), with \(a_2 > 0\), \(b_2 \ge 0\), and \(\lambda_2 \ge 0\).

theta

Clayton copula parameter with \(\theta >0\).

Value

A list containing:

(U, V)

Independent generated uniform numbers for \((X,Y)\).

(X, Y)

Dependent generated \(n\) pairs \((X,Y)\) observations.

Details

Further details are provided in Kızılaslan (2026).

References

Kızılaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv:????

Examples

set.seed(123)
n <- 50
a1 <- 0.75; b1 <- 1.5; lambda1 <- 0.6
a2 <- 1.2; b2 <- 0.5; lambda2 <- 0.9
theta <- 1 # 2, 3, 4, 5
# data generation
dat <- SSReliabilityClaytonMWD::rMweibull_Clayton(n, a1, b1, lambda1, a2, b2, lambda2, theta)
str(dat)
#> List of 4
#>  $ U: num [1:50] 0.288 0.788 0.409 0.883 0.94 ...
#>  $ V: num [1:50] 0.0726 0.6101 0.7749 0.3214 0.7372 ...
#>  $ X: num [1:50] 0.485 1.062 0.617 1.231 1.388 ...
#>  $ Y: num [1:50] 0.00392 0.33629 0.56182 0.08896 0.50229 ...