Estimating parameters of SSR Model with MWD Marginals via Clayton Copula

Estimates the parameters of all model parameters such as \((a_1, b_1,\lambda_1)\) for the strength variable \(X\), \((a_2, b_2, \lambda_2)\) for the stress variable \(Y\), and \(\theta\) for the Clayton copula parameter.

Fits a dependent stress–strength reliability (SSR) model in which both stress and strength marginal distributions follow the Modified Weibull Distribution (MWD), and dependence is modeled using a Clayton copula.

The function estimates marginal and copula parameters using several classical methods, including Maximum Likelihood Estimation (MLE), Least Squares Estimation (LSE), Weighted Least Squares Estimation (WLSE), and Maximum Product of Spacings (MPS).

Usage

fit.SSR.ClaytonMWD(
  data,
  ACI = FALSE,
  bootstrap = FALSE,
  B = NULL,
  seed = NULL,
  one.step = TRUE,
  alpha = 0.05,
  verbose = FALSE
)

Arguments

data

A list with two numeric vectors: \(X\) (strength) and \(Y\) (stress).

ACI

Logical; if TRUE, \(95\%\) asymptotic confidence intervals based on MLEs are computed.

bootstrap

Logical; if TRUE, \(95\%\) parametric bootstrap intervals are computed.

B

Number of bootstrap samples.

seed

Integer seed for reproducibility.

one.step

Logical; if TRUE, one-step LSE and WLSE methods are used for estimating \(\theta\).

alpha

Numeric; significance level for confidence intervals (e.g., \(0.05\) for \(95\%\) CI).

verbose

Logical; if TRUE, progress and intermediate results from the optimization procedure are printed. Default is FALSE.

Value

A list containing:

all.results

Point estimates of all model parameters.

theta.Ktau

Kendall's Tau estimate of \(theta\).

seed

Used seed in the analysis.

data

Used data in the analysis.

ACI.parameters

If ACI=TRUE, it represents the lower and upper bounds of the \(100(1-\alpha)%\) ACI of the parameters with length.

boot.mle

If bootstrap=TRUE, it represents the lower and upper bounds of the \(100(1-\alpha)%\) parametric bootstrap intervals of the parameters based on MLE.

boot.lse

If bootstrap=TRUE, it represents the lower and upper bounds of the \(100(1-\alpha)%\) parametric bootstrap intervals of the parameters based on LSE.

boot.wlse

If bootstrap=TRUE, it represents the lower and upper bounds of the \(100(1-\alpha)%\) parametric bootstrap intervals of the parameters based on WLSE.

boot.mps

If bootstrap=TRUE, it represents the lower and upper bounds of the \(100(1-\alpha)%\) parametric bootstrap intervals of the parameters based on MPS.

boot.samples

A list containing of all the bootstrap samples acrross all the parameters and methods.

Details

Fitting SSR Model with MWD Marginals via Clayton Copula

Returns point and interval estimates of model parameters using MLE, LSE, WLSE, and MPS methods. Further details can be found in Kızılaslan (2026).

References

Kızılaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv:????

Examples

data = list(X = TerkosDam, Y = OmerliDam)
fit.SSR = fit.SSR.ClaytonMWD(data,  ACI = TRUE, bootstrap = FALSE, B = 1000,
                             seed = 2026, one.step = TRUE, alpha = 0.05)
print(fit.SSR)
#> 
#> Fitting Results of Stress-Strength Reliability Model with MWD Marginals via Clayton Copula 
#> 
#> 
#> Table: Parameter Estimates
#> 
#> |     |      a1|      b1| lambda1|      a2|      b2| lambda2|   theta|       R|
#> |:----|-------:|-------:|-------:|-------:|-------:|-------:|-------:|-------:|
#> |mle  | 0.98542| 2.66095| 2.11322| 0.02428| 0.45908| 6.33059| 0.50551| 0.50428|
#> |lse  | 0.07196| 1.72091| 5.71667| 0.01166| 0.00001| 7.06623| 1.09884| 0.49349|
#> |wlse | 0.04261| 1.29079| 6.20629| 0.01154| 0.00001| 7.10656| 1.23631| 0.49824|
#> |mps  | 0.68997| 2.31111| 2.39000| 0.02043| 0.26258| 6.43699| 0.92547| 0.51110|
#> 
#> 
#> Table: 95% Asymptotic Confidence Intervals (MLE)
#> 
#> |       |      a1|      b1| lambda1|      a2|      b2| lambda2|   theta|       R|
#> |:------|-------:|-------:|-------:|-------:|-------:|-------:|-------:|-------:|
#> |lower  | 0.00000| 0.43992| 0.00000| 0.00000| 0.00000| 4.21722| 0.14734| 0.40310|
#> |upper  | 4.33682| 4.88197| 5.79639| 0.06723| 1.39756| 8.44395| 0.86369| 0.60545|
#> |length | 4.33682| 4.44205| 5.79639| 0.06723| 1.39756| 4.22673| 0.71636| 0.20235|
#> 
#> Kendall's tau estimate for theta: 1.25912