Estimating parameters of the two-parameter Weibull distribution

Estimates the parameters of the two-parameter Weibull distribution using classical methods.

fitWD(
  data,
  est.method,
  opt.method,
  starts,
  lower = NULL,
  upper = NULL,
  verbose = FALSE,
  ...
)

Arguments

data

Vector of observations.

est.method

Used method for estimating the parameters such as maximum likelihood estimate "MLE", the least square estimation "LSE", the weighted least square estimation "WLSE", and the maximum product of spacing estimates "MPS".

opt.method

The optimization method for optim function such as "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.

starts

Initial values for the parameters to be optimized over.

lower

Numeric vector specifying the lower bounds of the parameters for bounded optimization methods (e.g., "L-BFGS-B"). Ignored if NULL.

upper

Numeric vector specifying the upper bounds of the parameters for bounded optimization methods. Ignored if NULL.

verbose

Logical; if TRUE, progress and intermediate results from the optimization procedure are printed. Default is FALSE.

...

Further arguments to be passed to optim function such as lower and upper limits, hessian, etc.

Value

A list containing:

estimates

Estimated values of the model parameters.

measures

Model selection criteria, including the log-likelihood, AIC, and BIC, evaluated at the estimated parameter values.

initials

Initial values used in the optimization procedure.

opt.fit

Full optimization output.

Details

Classical estimations fo the parameters of the two-parameter Weibull distribution

The two-parameter Weibull Distribution has cumulative distribution function (CDF) and probability density function (PDF) given by

$$F(x) = 1 - \exp(-a x^b),$$ $$f(x) = a b x^{b - 1} \exp(-a x^b ),$$

where \(x > 0\), \(a > 0\) is the scale parameter and \(b > 0\) is the shape parameter.

The model parameters are estimated using several classical methods:

Maximum Likelihood Estimation (MLE): Obtains parameter estimates by maximizing the log-likelihood function of the observed data under the assumed two-parameter Weibull distribution.

Least Squares Estimation (LSE): Estimates parameters by minimizing the sum of squared differences between the empirical distribution function and theoretical distribution function of MWD. The Benard’s approximation is used for the empirical distribution function at the ordered observations as \(F(x_(i)) = (i-0.3)/(n+0.4), i=1,...,n.\).

Weighted Least Squares Estimation (WLSE): A modification of LSE that assigns weights to the squared differences. The weights are taken as \(w_i = ((n+1)^2(n+2)) / (i(n-i+1)), i=1,...,n. \)

Maximum Product of Spacings (MPS): Estimates parameters by maximizing the product of spacings between consecutive values of the fitted distribution function, providing a robust alternative to MLE, particularly in small samples.

Examples

# generate data from WD(a, b)
n <- 50
a <- 0.75; b <- 1.25; lambda <- 0 # reduces two-parameter Weibull distribution
set.seed(123)
X <- rMweibull(n, a, b, lambda)
init <- runif(2)

# fit model
fit.mle <- fitWD(data = X, est.method = "mle", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.mle$estimates
#>        a        b 
#> 0.675279 1.283785 

fit.lse <- fitWD(data = X, est.method = "lse", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.lse$estimates
#>         a         b 
#> 0.7012337 1.2147202 

fit.wlse <- fitWD(data = X, est.method = "wlse", opt.method = "L-BFGS-B", starts = init,
                  lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.wlse$estimates
#>        a        b 
#> 0.690969 1.236661 

fit.mps <- fitWD(data = X, est.method = "mps", opt.method = "L-BFGS-B", starts = init,
                 lower = c(1e-05,1e-05), upper = c(Inf,Inf), hessian = FALSE )
fit.mps$estimates
#>         a         b 
#> 0.6848478 1.1977846