Parametric Bootstrap Confidence Intervals

It gives the parametric bootstrap confidence interval estimation for all the unknown parameters along with the reliability \(R\) based on MLE, LSE, WLSE and MPS methods.

parametric_bootstrap(
  est.method,
  opt.method,
  boot.estimates,
  n,
  B = 1000,
  seed = NULL,
  one.step = TRUE,
  alpha = 0.05
)

Arguments

est.method

Used method for estimating the parameters such as maximum likelihood estimate "MLE", the least square estimation "LSE", the weighted least square estimation "WLSE", and the maximum product of spacing estimates "MPS".

opt.method

The optimization method for optim function such as "Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN" and "Brent" to be used for estimating the parameters.

boot.estimates

A list containing the estimates of the unknown parameters. The elements \((a_1, b_1, \lambda_1)\) correspond to the strength variable, \((a_2, b_2, \lambda_2)\) correspond to the stress variable, and \(\theta\) corresponds to Clayton copula parameter.

n

sample size

B

Number of bootstrap samples.

seed

Integer seed for reproducibility.

one.step

Logical; if TRUE, one-step LSE and WLSE methods are used for estimating \(\theta\).

alpha

Numeric; significance level for confidence intervals (e.g., \(0.05\) for \(95\%\) CI).

Value

A list containing:

parameters.quantiles

It gives the \(100(1-\alpha)\%\) lower and upper bounds.

boot.results

It represents all the estimates values of the parameters through B bootstrap samples.

Details

The parametric bootstrap percentile method is employed to be able to construct bootstrap confidence intervals for unknown parameters and \(R\) as well based on MLE, LSE, WLSE and MPSE methods. Further details can be found in Kızılaslan (2026).

References

Kizilaslan, F. (2026). Reliability estimation in dependent stress–strength model with Clayton copula and modified Weibull margins. arXiv:2604.12130